requirements.txt
    click==7.1.2 
    numpy==1.21.4 
    pandas==1.1.4 
    pillow==8.0.0 
    scikit-learn==0.23.2 
    torch==1.11.0 
    torchvision==0.12.0
